requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
boto3==1.40.6
botocore==1.40.6
click==8.2.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
jmespath==1.0.1
mako==1.3.10
markupsafe==3.0.2
passlib==1.7.4
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.42
starlette==0.47.2
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
